backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai
beautifulsoup4
requests
selenium
webdriver-manager
